requirements.txt
numpy
pandas
scikit-learn
yfinance

tensorflow
keras

transformers
torch
sentencepiece

chromadb
sentence-transformers

fastapi
uvicorn
requests
pydantic

anthropic
python-dotenv

streamlit
plotly